Pull some functionality to AbstractOfflineManager
